+++ web/backend/service/app.py
@@ -13,11 +13,9 @@ class App:
         self._memory = sc.Memory(self._config.get_path_to_sc_config())
 
     def run(self):
-        from .flask_app import init_app_from_config
-        from .flask_app.db import db
+        from .flask_app import create_app
 
-        app = init_app_from_config(self._config)
-        db.init_app(app)
+        app = create_app(self._config)
 
         try:
             app.run(self._config.get_host(), self._config.get_port())
