+++ web/backend/service/config.py
@@ -1,9 +1,6 @@
 import configparser
 import os
 
-from typing import KeysView
-from secrets import token_hex
-
 
 class Config:
 
@@ -19,7 +16,7 @@ class Config:
             return default
 
     def get_secret(self) -> str:
-        return self._get_safe_value("flask", "secret", token_hex(32))
+        return self._get_safe_value("flask", "secret", "dev")
 
     def get_host(self) -> str:
         return self._get_safe_value("flask", "host", "127.0.0.1")
@@ -27,6 +24,8 @@ class Config:
     def get_port(self) -> int:
         return int(self._get_safe_value("flask", "port", "5000"))
 
+    def get_db_address(self) -> str:
+        return self._get_safe_value("flask", "sql", "sqlite:///../data.sqlite")
 
     def get_path_to_sc_config(self) -> str:
         path = self._get_safe_value("sc", "config", "")
